# docs/openapi.yaml
openapi: 3.0.0
info:
  title: SmartRetailX Microservices API
  description: Complete API specifications contract for the cloud-native SmartRetailX e-commerce platform. Covers user auth, catalog queries, order placement, payments, and inventory warehouse telemetry.
  version: 1.0.0
servers:
  - url: http://abc3dc5c1549c4f829ade5d0c65a0af6-939735814.us-east-1.elb.amazonaws.com:8000/api/v1
    description: Production EKS API Gateway LoadBalancer
  - url: http://localhost:8000/api/v1
    description: Local Docker Compose Gateway Ingress
paths:
...
  /products:
    get:
      summary: Retrieve all product catalog collections
      description: Returns a cached list of active shopping products.
      responses:
        '200':
          description: A list of products.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
    post:
      summary: Create new catalog item
      description: Adds a new product configuration to MongoDB and flushes Redis cache. Restricted to admins/managers.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        '201':
          description: Product created successfully.
        '403':
          description: Forbidden. Insufficient RBAC permission.
  /products/{product_id}:
    get:
      summary: Get product details
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
    delete:
      summary: Delete product from catalog
      security:
        - bearerAuth: []
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product deleted successfully.
        '403':
          description: Forbidden. Insufficient RBAC permission.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Product:
      type: object
      properties:
        product_id:
          type: string
          example: PRD-001
        name:
          type: string
          example: Smart Wireless Earbuds
        category:
          type: string
          example: Electronics
        price:
          type: number
          example: 79.99
        description:
          type: string
          example: True wireless earbuds with active noise cancellation.
    ProductInput:
      type: object
      required:
        - name
        - category
        - price
      properties:
        name:
          type: string
          example: Mechanical Keyboard
        category:
          type: string
          example: Electronics
        price:
          type: number
          example: 120.00
        description:
          type: string
          example: Tactile blue switch keyboard with RGB backlights.
...